Refactoring some e2e tests.  Now is use a global-setup instead of repeat repeat code to setup tests enviroment.
